requirements.txt
icalendar
requests
bs4
fastapi
pydantic
mysql-connector-python
uvicorn
aiofiles
pycryptodome
ddddocr
